WIP: Expose test tools as features
