+++ crates/dbcrossbar_trino_types/Cargo.toml
@@ -3,25 +3,55 @@ name = "dbcrossbar_trino_types"
 version = "0.1.0"
 edition = "2021"
 
+# Configuration for docs.rs to enable all features and annotate APIs with the
+# feature they are defined in.
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
+[features]
+default = []
+values = [
+    "dep:base64",
+    "dep:chrono",
+    "dep:float-cmp",
+    "dep:geojson",
+    "dep:geo-types",
+    "dep:rust_decimal",
+    "dep:serde_json",
+    "dep:uuid",
+    "dep:wkt",
+]
+proptest = [
+    "values",
+    "dep:proptest",
+    "dep:proptest-arbitrary-interop",
+    "dep:proptest-derive",
+]
+client = ["values", "dep:reqwest", "dep:tokio"]
+
 [dev-dependencies]
-base64 = "0.22.1"
-chrono = { version = "0.4.38", features = ["arbitrary"] }
-float-cmp = "0.10.0"
-geojson = { version = "0.24.1", features = ["geo-types"] }
-geo-types = "0.7.13"
-proptest = "1.5.0"
-proptest-arbitrary-interop = "0.1.0"
-proptest-derive = "0.5.0"
-prusto = "0.5.1"
+proptest = { version = "1.5.0" }
+proptest-arbitrary-interop = { version = "0.1.0" }
+proptest-derive = { version = "0.5.0" }
+tokio = { version = "1.16", features = ["macros", "rt-multi-thread", "time"] }
+
+[dependencies]
+base64 = { version = "0.22.1", optional = true }
+chrono = { version = "0.4.38", features = ["arbitrary"], optional = true }
+float-cmp = { version = "0.10.0", optional = true }
+geojson = { version = "0.24.1", features = ["geo-types"], optional = true }
+geo-types = { version = "0.7.13", optional = true }
+proptest = { version = "1.5.0", optional = true }
+proptest-arbitrary-interop = { version = "0.1.0", optional = true }
+proptest-derive = { version = "0.5.0", optional = true }
 reqwest = { version = "0.12.8", default-features = false, features = [
     "http2",
     "json",
-] }
-round = "0.1.2"
+], optional = true }
+rust_decimal = { version = "1.36.0", optional = true }
 serde = { version = "1.0.210", features = [] }
-serde_json = "1.0.128"
-tokio = { version = "1.16", features = ["time"] }
-uuid = { version = "1.10.0", features = ["arbitrary"] }
-wkt = "0.11.1"
-
-[dependencies]
+serde_json = { version = "1.0.128", optional = true }
+tokio = { version = "1.16", features = ["macros", "time"], optional = true }
+uuid = { version = "1.10.0", features = ["arbitrary"], optional = true }
+wkt = { version = "0.11.1", optional = true }
