+++ crates/dbcrossbar_trino_types/src/client/errors.rs
